// src/types.ts
export type FilterValueType = "Inflation" | "Natural disasters" | "Crises" | "Geopolitical factors";

export type Level = 1 | 2 | 3 | 4 | 5;

export enum Section {
  STATE = "STATE",
  SOCIAL = "SOCIAL",
  EDU = "EDU",
  GOV = "GOV",
  SOC_1 = "SOC_1",
  EXT = "EXT",
  INT = "INT",
  PRIVATE = "PRIVATE",
}